(refactor) Put VRScene into new PhodomeScene component
